requirements.txt
aniso8601==1.2.0
blinker==1.4
boto3==1.4.4
botocore==1.5.33
click==6.6
docutils==0.13.1
Flask==0.11.1
Flask-Cors==3.0.2
Flask-JWT==0.3.2
Flask-PyMongo==0.4.1
Flask-RESTful==0.3.5
gunicorn==19.7.1
idna==2.1
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.2
MarkupSafe==0.23
numpy==1.11.2
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
pymongo==3.3.1
python-dateutil==2.6.0
pytz==2016.7
s3transfer==0.1.10
six==1.10.0
Werkzeug==0.11.11
wheel==0.24.0